[OpenMP][NFC] Put ExponentialBackoff in a Utils header
"private.h" will go.
